[TASK] Move syncPage into separate folder
